requirements.txt
Django==3.0.6
django-bootstrap3==12.1.0
django-material==1.6.7
django-tinymce==3.0.2
djangorestframework==3.11.0
numpy==1.8.2
Pillow==7.1.2
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1